requirements.txt
Django==2.1.5
django-ckeditor==5.6.1
django-js-asset==1.1.0
psycopg2-binary==2.7.7
pytz==2018.9
